docs+test: P2SP 阶段 0 批 4 收尾——swarmd_config 补矿 14 用例 + MSVC 编译面两修复 + 三份文档 + 铁账
- swarmd_config.cpp 从可执行移入 falcon_swarm_server（PRIVATE 链 falcon_daemon_core）+ 新增 swarm_config_test.cpp 14 用例（往返/六类错误路径/告警/~/展开）——config 矿净收 94 行，测试 70 → 84
- MSVC 修复（批 3 run 36076043736 Qt6 Windows 唯一红）：chrono rep 窄化 C2397 显式 static_cast；POSIX ::poll C3861 → #define poll WSAPoll
- 铁账：分母 19580、miss 890、行 95.5%、函数 98.5%、分支 55.0%（新包增量稀释，config 可测面收口）
- 三份 CLAUDE.md（根批 4 条目 + daemon 拆分 + swarmd 包文档）

+++ packages/falcon-swarmd/src/client/swarm_client.cpp
@@ -29,7 +29,10 @@ SwarmClient::SwarmClient(SwarmClientConfig cfg, SwarmKeyMaterial key)
     : cfg_(std::move(cfg)),
       key_(std::move(key)),
       http_(SwarmHttpClient::Options{cfg_.host, cfg_.port, cfg_.server_token,
-                                     cfg_.rpc_timeout_ms.count()}) {}
+                                     // chrono rep 在 MSVC 是 long long，
+                                     // brace-init 到 long 需显式收窄
+                                     static_cast<long>(
+                                         cfg_.rpc_timeout_ms.count())}) {}
 
 SwarmClient::~SwarmClient() { stop(); }
